Update noon forecast table to include updateDate for new records
