+++ api/alembic/versions/10560b37a029_forecast_model_update.py
@@ -1,0 +1,43 @@
+"""Forecast model update
+
+Revision ID: 10560b37a029
+Revises: 1caf3488a340
+Create Date: 2021-12-02 14:03:38.474745
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '10560b37a029'
+down_revision = '1caf3488a340'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('noon_forecasts', sa.Column('update_date', sa.TIMESTAMP(timezone=True), nullable=True))
+    op.alter_column('noon_forecasts', 'danger_rating',
+                    existing_type=sa.INTEGER(),
+                    nullable=True)
+    op.drop_constraint('noon_forecasts_weather_date_station_code_temp_valid_tempera_key',
+                       'noon_forecasts', type_='unique')
+    op.create_index(op.f('ix_noon_forecasts_update_date'), 'noon_forecasts', ['update_date'], unique=False)
+    op.create_unique_constraint('ix_unique_constraint', 'noon_forecasts', [
+                                'weather_date', 'station_code', 'update_date'])
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint('ix_unique_constraint', 'noon_forecasts', type_='unique')
+    op.drop_index(op.f('ix_noon_forecasts_update_date'), table_name='noon_forecasts')
+    op.create_unique_constraint('noon_forecasts_weather_date_station_code_temp_valid_tempera_key', 'noon_forecasts', [
+                                'weather_date', 'station_code', 'temp_valid', 'temperature', 'rh_valid', 'relative_humidity', 'wdir_valid', 'wind_direction', 'wspeed_valid', 'wind_speed', 'precip_valid', 'precipitation', 'gc', 'ffmc', 'dmc', 'dc', 'isi', 'bui', 'fwi', 'danger_rating'])
+    op.alter_column('noon_forecasts', 'danger_rating',
+                    existing_type=sa.INTEGER(),
+                    nullable=False)
+    op.drop_column('noon_forecasts', 'update_date')
+    # ### end Alembic commands ###
