--- api/alembic/versions/df329df5bc70_forecast_model_has_default_math_nan_.py
@@ -1,32 +1,0 @@
-"""Forecast model has default math.nan values
-
-Revision ID: df329df5bc70
-Revises: 1caf3488a340
-Create Date: 2021-12-02 11:50:15.785584
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'df329df5bc70'
-down_revision = '1caf3488a340'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('noon_forecasts', 'danger_rating',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('noon_forecasts', 'danger_rating',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    # ### end Alembic commands ###
